% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cleandata_2011}
\alias{cleandata_2011}
\title{Significant earthquakes around the world - cleaned and filtered for 2011}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 59 rows and 48 columns.}
\source{
Obtained from the U.S. National Oceanographic and Atmospheric Administration (NOAA), , cleaned and filtered.
}
\usage{
cleandata_2011
}
\description{
Significant earthquakes around the world - cleaned and filtered for 2011
}
\examples{
\dontrun{
cleandata_2011
}

}
\keyword{datasets}
