% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cleandata_MEXICO2000}
\alias{cleandata_MEXICO2000}
\title{Significant earthquakes around the world - cleaned and filtered for MEXICO 2000}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24 rows and 48 columns.}
\source{
Obtained from the U.S. National Oceanographic and Atmospheric Administration (NOAA), cleaned and filtered.
}
\usage{
cleandata_MEXICO2000
}
\description{
Significant earthquakes around the world - cleaned and filtered for MEXICO 2000
}
\examples{
\dontrun{
cleandata_MEXICO2000
}

}
\keyword{datasets}
