% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleandata.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean the raw data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{A raw NOAA dataset as a dataframe that has been already read into R}
}
\value{
a clean dataframe with new date column
}
\description{
This is a function that reads a raw NOAA data file and creates date column
by uniting the year, month, day and converting it to the Date class.
Also, it converts LATITUDE and LONGITUDE columns to numeric class
}
\note{
the function does not read data from a file on a disk
}
\examples{
\dontrun{
eq_clean_data("rawdata")
}

}
