% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletmap.R
\name{eq_map}
\alias{eq_map}
\title{Mapping the earthquake epicenters}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{data containing the filtered data frame with earthquakes to visualize}

\item{annot_col}{column used for the annotation in the pop-up}
}
\value{
A map
}
\description{
Mapping the earthquake epicenters and providing some annotations with the mapped data.
Takes an argument data containing the filtered data frame with earthquakes to visualize.
The function maps the epicenters (LATITUDE/LONGITUDE) and annotates each point with in
pop up window containing annotation data stored in a column of the data frame.
The user can choose which column is used for the annotation in the pop-up with
a function argument named annot_col. Each earthquake is shown with a circle,
and the radius of the circle is proportional to the earthquake's magnitude (EQ_PRIMARY).
}
\examples{
\dontrun{
readr::read_delim("Data/earthquakes.txt", delim = "\\t") \%>\%
                  eq_clean_data() \%>\%
                  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
                  eq_map(annot_col = "DATE")
}

}
