% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rawdata}
\alias{rawdata}
\title{Significant earthquakes around the world}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6083 rows and 47 columns.}
\source{
Obtained from the U.S. National Oceanographic and Atmospheric Administration (NOAA)
}
\usage{
rawdata
}
\description{
Significant earthquakes around the world
}
\examples{
\dontrun{
rawdata
}

}
\keyword{datasets}
